refactor : delete memberService, add code in authController
