+++ src/main/java/com/daangn/auth/controller/AuthController.java
@@ -3,6 +3,8 @@ package com.daangn.auth.controller;
 import com.daangn.auth.dto.LoginRequestDTO;
 import com.daangn.auth.dto.TokenResponseDTO;
 import com.daangn.auth.service.AuthService;
+import com.daangn.member.domain.Member;
+import com.daangn.member.dto.MemberDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -23,6 +25,13 @@ public class AuthController {
     public ResponseEntity<TokenResponseDTO> login(@RequestBody @Valid LoginRequestDTO loginRequestDTO) {
         return ResponseEntity.ok()
                 .body(authService.login(loginRequestDTO));
+    }
+
+    @PostMapping("/signup")
+    public ResponseEntity<Member> createUser(@RequestBody MemberDto member) {
+        authService.existsEmail(member.getEmail());
+        return ResponseEntity.ok()
+                .body(authService.createMember(member));
     }
 
 }
